Adding Discount class as a Base Class
Discount class will implement IDiscount interface
RelativeDiscount derived from discount class
This is because of having new type of discount the UPC discount
