+++ PriceCalculatorKata/src/PriceCalculator/RelativeDiscount.cs
@@ -1,6 +1,6 @@
 namespace PriceCalculator;
 
-public class RelativeDiscount : IDiscount
+public class RelativeDiscount : Discount, IDiscount
 {
     private static readonly RelativeDiscount s_discountInstance = new();
 
@@ -8,25 +8,8 @@ public class RelativeDiscount : IDiscount
     {
     }
 
-    public int DiscountValue { get; private set; }
-
     public static RelativeDiscount GetDiscountInstance()
     {
         return s_discountInstance;
-    }
-
-    public void SetDiscount(string discount)
-    {
-        if (isValidDiscount(discount))
-            DiscountValue = int.Parse(discount);
-    }
-
-    private bool isValidDiscount(string discount)
-    {
-        int result;
-        if (string.IsNullOrWhiteSpace(discount)) return false;
-        if (!int.TryParse(discount, out result)) return false;
-        if (result < 0) return false;
-        return true;
     }
 }
